move gasTracker tests files under __tests__ folder
